[EasyWebhook] Make sure webhook is persisted before send async
