+++ packages/EasyWebhook/src/Exceptions/WebhookIdRequiredForAsyncException.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace EonX\EasyWebhook\Exceptions;
+
+final class WebhookIdRequiredForAsyncException extends AbstractEasyWebhookException
+{
+    // No body needed.
+}
